split command_queue_interface.hpp into pieces
